Implement lazy-loading / MVC in results section
