+++ setup.cfg
@@ -57,8 +57,8 @@ aiidalab_qe.plugins.xas = pseudo_toc.yaml
 aiidalab_qe.properties =
     bands = aiidalab_qe.plugins.bands:bands
     pdos = aiidalab_qe.plugins.pdos:pdos
-    xps = aiidalab_qe.plugins.xps:xps
     electronic_structure = aiidalab_qe.plugins.electronic_structure:electronic_structure
+    xps = aiidalab_qe.plugins.xps:xps
     xas = aiidalab_qe.plugins.xas:xas
 
 aiida.workflows =
