+++ src/aiidalab_qe/app/result/summary/__init__.py
@@ -1,0 +1,7 @@
+from .model import WorkChainSummaryModel
+from .summary import WorkChainSummary
+
+__all__ = [
+    "WorkChainSummaryModel",
+    "WorkChainSummary",
+]
