+++ src/aiidalab_qe/app/result/viewer/__init__.py
@@ -1,0 +1,7 @@
+from .model import WorkChainViewerModel
+from .viewer import WorkChainViewer
+
+__all__ = [
+    "WorkChainViewerModel",
+    "WorkChainViewer",
+]
