+++ src/aiidalab_qe/common/bands_pdos/__init__.py
@@ -1,0 +1,7 @@
+from .bandpdoswidget import BandPdosWidget
+from .model import BandsPdosModel
+
+__all__ = [
+    "BandPdosWidget",
+    "BandsPdosModel",
+]
