+++ src/aiidalab_qe/common/bands_pdos/model.py
@@ -1,0 +1,150 @@
+import base64
+import json
+
+import ipywidgets as ipw
+import numpy as np
+import traitlets as tl
+from IPython.display import display
+
+from aiida.common.extendeddicts import AttributeDict
+from aiidalab_qe.common.bands_pdos.utils import (
+    get_bands_projections_data,
+    get_pdos_data,
+)
+from aiidalab_qe.common.mvc import Model
+from aiidalab_widgets_base.utils import string_range_to_list
+
+
+class BandsPdosModel(Model):
+    bands = tl.Instance(AttributeDict, allow_none=True)
+    pdos = tl.Instance(AttributeDict, allow_none=True)
+
+    dos_atoms_group_options = tl.List(
+        trait=tl.List(tl.Unicode()),
+        default_value=[
+            ("Kinds", "kinds"),
+            ("Atomic position", "atoms"),
+        ],
+    )
+    dos_atoms_group = tl.Unicode("kinds")
+    dos_plot_group_options = tl.List(
+        trait=tl.List(tl.Unicode()),
+        default_value=[
+            ("Total", "total"),
+            ("Orbital", "orbital"),
+            ("Angular momentum", "angular_momentum"),
+        ],
+    )
+    dos_plot_group = tl.Unicode("total")
+    selected_atoms = tl.Unicode("")
+    project_bands_box = tl.Bool(False)
+    proj_bands_width = tl.Float(0.5)
+
+    pdos_data = tl.Dict()
+    bands_data = tl.Dict()
+    needs_pdos_options = tl.Bool(False)
+    needs_projections_controls = tl.Bool(False)
+
+    ready = tl.Bool(False)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        ipw.dlink(
+            (self, "bands"),
+            (self, "needs_projections_controls"),
+            lambda node: node is not None and "projwfc" in node,
+        )
+        ipw.dlink(
+            (self, "pdos"),
+            (self, "needs_pdos_options"),
+            lambda _: self.pdos is not None or self.needs_projections_controls,
+        )
+        ipw.dlink(
+            (self, "needs_projections_controls"),
+            (self, "needs_pdos_options"),
+            lambda _: self.pdos is not None or self.needs_projections_controls,
+        )
+        ipw.dlink(
+            (self, "bands"),
+            (self, "ready"),
+            lambda node: bool(node),
+        )
+        ipw.dlink(
+            (self, "pdos"),
+            (self, "ready"),
+            lambda node: bool(node),
+        )
+
+    def fetch_data(self):
+        """Fetch the data from the nodes."""
+        if self.bands:
+            self.bands_data = self._get_bands_data()
+        if self.pdos:
+            self.pdos_data = self._get_pdos_data()
+
+    def _get_pdos_data(self):
+        if not self.pdos:
+            return None
+        expanded_selection, syntax_ok = string_range_to_list(
+            self.selected_atoms, shift=-1
+        )
+        if syntax_ok:
+            return get_pdos_data(
+                self.pdos,
+                group_tag=self.dos_atoms_group,
+                plot_tag=self.dos_plot_group,
+                selected_atoms=expanded_selection,
+            )
+        return None
+
+    def _get_bands_data(self):
+        if not self.bands:
+            return None
+
+        expanded_selection, syntax_ok = string_range_to_list(
+            self.selected_atoms, shift=-1
+        )
+        if syntax_ok:
+            return get_bands_projections_data(
+                self.bands,
+                group_tag=self.dos_atoms_group,
+                plot_tag=self.dos_plot_group,
+                selected_atoms=expanded_selection,
+                bands_width=self.proj_bands_width,
+            )
+        return None
+
+    def download_data(self, _=None):
+        """Function to download the data."""
+        if self.bands_data:
+            bands_data_export = {
+                key: value.tolist() if isinstance(value, np.ndarray) else value
+                for key, value in self.bands_data.items()
+            }
+            json_str = json.dumps(bands_data_export)
+            b64_str = base64.b64encode(json_str.encode()).decode()
+            file_name_bands = "bands_data.json"
+            self._download(payload=b64_str, filename=file_name_bands)
+        if self.pdos_data:
+            json_str = json.dumps(self.pdos_data)
+            b64_str = base64.b64encode(json_str.encode()).decode()
+            file_name_pdos = "dos_data.json"
+            self._download(payload=b64_str, filename=file_name_pdos)
+
+    @staticmethod
+    def _download(payload, filename):
+        """Download payload as a file named as filename."""
+        from IPython.display import Javascript
+
+        javas = Javascript(
+            f"""
+            var link = document.createElement('a');
+            link.href = 'data:text/json;charset=utf-8;base64,{payload}'
+            link.download = "{filename}"
+            document.body.appendChild(link);
+            link.click();
+            document.body.removeChild(link);
+            """
+        )
+        display(javas)
