+++ src/aiidalab_qe/plugins/bands/result/__init__.py
@@ -1,0 +1,7 @@
+from .model import BandsResultModel
+from .result import BandsResult
+
+__all__ = [
+    "BandsResultModel",
+    "BandsResult",
+]
