+++ src/aiidalab_qe/plugins/electronic_structure/result/__init__.py
@@ -1,0 +1,7 @@
+from .model import ElectronicStructureResultModel
+from .result import ElectronicStructureResult
+
+__all__ = [
+    "ElectronicStructureResultModel",
+    "ElectronicStructureResult",
+]
