+++ src/aiidalab_qe/plugins/pdos/result/__init__.py
@@ -1,0 +1,7 @@
+from .model import PdosResultModel
+from .result import PdosResult
+
+__all__ = [
+    "PdosResultModel",
+    "PdosResult",
+]
