+++ src/aiidalab_qe/plugins/xas/result/__init__.py
@@ -1,0 +1,7 @@
+from .model import XasResultModel
+from .result import XasResult
+
+__all__ = [
+    "XasResultModel",
+    "XasResult",
+]
