+++ src/aiidalab_qe/plugins/xps/result/__init__.py
@@ -1,0 +1,7 @@
+from .model import XpsResultModel
+from .result import XpsResult
+
+__all__ = [
+    "XpsResultModel",
+    "XpsResult",
+]
