+++ tests/test_plugins_pdos.py
@@ -1,25 +1,30 @@
 def test_result(generate_qeapp_workchain):
     import plotly.graph_objects as go
 
-    from aiidalab_qe.common.bandpdoswidget import BandPdosWidget
-    from aiidalab_qe.plugins.pdos.result import Result
+    from aiidalab_qe.common.bands_pdos import BandPdosWidget
+    from aiidalab_qe.plugins.pdos.result import PdosResult, PdosResultModel
 
-    wkchain = generate_qeapp_workchain()
+    workchain = generate_qeapp_workchain()
     # generate structure for scf calculation
-    result = Result(node=wkchain.node)
-    result._update_view()
-    assert isinstance(result.children[0], BandPdosWidget)
-    assert isinstance(result.children[0].bandsplot_widget, go.FigureWidget)
+    model = PdosResultModel()
+    model.process_node = workchain.node
+    result = PdosResult(model=model)
+    result.render()
+
+    widget = result.children[0]
+    model = widget._model
+
+    assert isinstance(widget, BandPdosWidget)
+
+    widget.plot_button.click()
+    assert isinstance(widget.plot, go.FigureWidget)
 
     # Check if data is correct
-    assert result.children[0].bands_data is None
-    assert result.children[0].pdos_data is not None
+    assert not model.bands_data
+    assert model.pdos_data
 
     # Check PDOS settings is not None
 
     # Check Bands axis
-    assert (
-        result.children[0].bandsplot_widget.layout.xaxis.title.text
-        == "Density of states (eV)"
-    )
-    assert result.children[0].bandsplot_widget.layout.yaxis.title.text is None
+    assert widget.plot.layout.xaxis.title.text == "Density of states (eV)"
+    assert widget.plot.layout.yaxis.title.text is None
