+++ tests/test_result.py
@@ -1,4 +1,10 @@
+import time
+
+from bs4 import BeautifulSoup
+
 from aiidalab_qe.app.main import App
+from aiidalab_qe.app.result.summary import WorkChainSummary, WorkChainSummaryModel
+from aiidalab_qe.app.result.viewer import WorkChainViewer
 
 
 def test_result_step(app_to_submit, generate_qeapp_workchain):
@@ -6,7 +12,7 @@ def test_result_step(app_to_submit, generate_qeapp_workchain):
     is running."""
     app: App = app_to_submit
     step = app.results_step
-    app.results_model.process = generate_qeapp_workchain().node.uuid
+    app.results_model.process_uuid = generate_qeapp_workchain().node.uuid
     assert step.state == step.State.ACTIVE
 
 
@@ -16,59 +22,49 @@ def test_kill_and_clean_buttons(app_to_submit, generate_qeapp_workchain):
     step = app_to_submit.results_step
     step.render()
     model = app_to_submit.results_model
-    model.process = generate_qeapp_workchain().node.uuid
+    model.process_uuid = generate_qeapp_workchain().node.uuid
     assert step.kill_button.layout.display == "block"
     assert step.clean_scratch_button.layout.display == "none"
 
 
 def test_workchainview(generate_qeapp_workchain):
     """Test the result tabs are properly updated"""
-    import time
 
-    from aiidalab_qe.app.result.workchain_viewer import WorkChainViewer
-
-    wkchain = generate_qeapp_workchain()
-    wkchain.node.seal()
-    wcv = WorkChainViewer(wkchain.node)
+    workchain = generate_qeapp_workchain()
+    workchain.node.seal()
+    viewer = WorkChainViewer(workchain.node)
     time.sleep(3)
-    assert len(wcv.result_tabs.children) == 5
-    assert wcv.result_tabs._titles["0"] == "Workflow Summary"  # type: ignore
-    assert wcv.result_tabs._titles["1"] == "Final Geometry"  # type: ignore
+    assert len(viewer.tabs.children) == 5
+    assert viewer.tabs._titles["0"] == "Workflow Summary"  # type: ignore
+    assert viewer.tabs._titles["1"] == "Final Geometry"  # type: ignore
 
 
 def test_summary_report(data_regression, generate_qeapp_workchain):
     """Test the summary report can be properly generated."""
-    from aiidalab_qe.app.result.summary_viewer import SummaryView
-
-    wkchain = generate_qeapp_workchain()
-    viewer = SummaryView(wkchain.node)
-    report = viewer.report
-    # regression test
-    data_regression.check(report)
+    workchain = generate_qeapp_workchain()
+    model = WorkChainSummaryModel()
+    model.process_node = workchain.node
+    report_parameters = model._generate_report_parameters()
+    data_regression.check(report_parameters)
 
 
 def test_summary_report_advanced_settings(data_regression, generate_qeapp_workchain):
     """Test advanced settings are properly reported"""
-    from aiidalab_qe.app.result.summary_viewer import SummaryView
-
-    wkchain = generate_qeapp_workchain(
+    workchain = generate_qeapp_workchain(
         spin_type="collinear", electronic_type="metal", initial_magnetic_moments=0.1
     )
-    viewer = SummaryView(wkchain.node)
-    report = viewer.report
-    assert report["initial_magnetic_moments"]["Si"] == 0.1
+    model = WorkChainSummaryModel()
+    model.process_node = workchain.node
+    report_parameters = model._generate_report_parameters()
+    assert report_parameters["initial_magnetic_moments"]["Si"] == 0.1
 
 
 def test_summary_view(generate_qeapp_workchain):
     """Test the report html can be properly generated."""
-    from bs4 import BeautifulSoup
-
-    from aiidalab_qe.app.result.summary_viewer import SummaryView
-
-    wkchain = generate_qeapp_workchain()
-    viewer = SummaryView(wkchain.node)
-    report_html = viewer.report_html
-    # report_html = generate_report_html(wcv.node)
+    workchain = generate_qeapp_workchain()
+    model = WorkChainSummaryModel()
+    model.process_node = workchain.node
+    report_html = model.generate_report_html()
     parsed = BeautifulSoup(report_html, "html.parser")
     # find the td with the text "Initial Magnetic Moments"
     parameters = {
